Convert market and staff modules to comptroller string uid ids
Models, repositories, Retrofit APIs, viewmodels, navigation callbacks, and
screens move every row/FK id (listing, order, auction, thread, category,
dispute, review, appeal, event, asset, subscription) from Long/Int to String,
matching the comptroller migration to mochi.uid() text ids. Money, counts,
timestamps, and 0/1 flags stay numeric. Also flips the 10 id-bearing string
labels from %d to %s across all locale catalogs so a uid renders without a
format crash.

+++ apps/staff/src/main/java/org/mochios/staff/ui/disputes/DisputesViewModel.kt
@@ -184,7 +184,7 @@ class DisputesViewModel @Inject constructor(
         viewModelScope.launch {
             try {
                 repository.reviewDispute(
-                    id = current.id.toInt(),
+                    id = current.id,
                     status = status,
                     resolution = resolution.takeIf { it.isNotBlank() },
                     refundAmount = refundAmountMinor,
